Move `hasTaskCallExpression` to utils
